requirements.txt
./
attrs==24.2.0
Cartopy==0.21.1
certifi==2024.8.30
cftime==1.6.4.post1
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.1.0
contourpy==1.1.1
cycler==0.12.1
dask==2023.5.0
exceptiongroup==1.2.2
fiona==1.10.1
fonttools==4.54.1
fsspec==2024.10.0
geopandas==0.13.2
importlib-metadata==8.5.0
importlib-resources==6.4.5
iniconfig==2.0.0
jinja2==3.1.4
kiwisolver==1.4.7
locket==1.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
netCDF4==1.7.2
numpy==1.24.4
packaging==24.1
pandas==2.0.3
partd==1.4.1
pillow==10.4.0
plotly==5.24.1
pluggy==1.5.0
pyparsing==3.1.4
pyproj==3.5.0
pyshp==2.3.1
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
shapely==2.0.6
six==1.16.0
tomli==2.0.2
toolz==1.0.0
tzdata==2024.2
xarray==2023.1.0
zipp==3.20.2